requirements.txt
requests
python-dotenv
tenacity
psycopg2-binary
tqdm
pandas
pyarrow
